# additional_life-openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - User API
      summary: Get user by ID
      description: Retrieves a user by their unique identifier
      operationId: getUserById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: User found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
      - User API
      summary: Update an existing user
      description: Updates a user with the provided data
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: User updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
      - User API
      summary: Delete a user
      description: Deletes a user by their unique identifier
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: User deleted successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: User not found
...
  /api/v1/chat/{id}/mark-as-read/{userId}:
    put:
      tags:
      - chat-controller
      operationId: markChatAsRead
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/public/users:
    post:
      tags:
      - Public User API
      summary: Create a new user
      description: Creates a new user with the provided data
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: User already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/items:
    post:
      tags:
      - Item API
      summary: Create a new item
      description: Creates a new item with the provided data
      operationId: createItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemRequest"
        required: true
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "422":
          description: Item could not be processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
...
  /api/v1/users/email/{email}:
    get:
      tags:
      - User API
      summary: Get user by email
      description: Retrieves a user by their email address
      operationId: getUserByEmail
      parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: User found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/v1/public/items:
    get:
      tags:
      - Public Item API
      summary: Get paginated list of available items
      description: Retrieves a paginated list of items that are active and not taken
      operationId: getItems
      parameters:
      - name: query
        in: query
        description: Full-text search query
        required: false
        schema:
          type: string
      - name: category
        in: query
        description: Item category
        required: false
        schema:
          type: string
          format: uuid
      - name: condition
        in: query
        description: Item condition
        required: false
        schema:
          type: string
      - name: sortBy
        in: query
        description: Sort field
        required: false
        schema:
          type: string
          default: createdAt
        example: createdAt
      - name: sortDirection
        in: query
        description: Sort direction
        required: false
        schema:
          type: string
          enum:
          - asc
          - desc
        example: desc
      - name: page
        in: query
        description: Page number (zero-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          minimum: 1
      responses:
        "200":
          description: Available items retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageItemResponse"
...
  /api/v1/chat/{id}/unread-count/{userId}:
    get:
      tags:
      - chat-controller
      operationId: getUnreadChatCount
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
...
components:
  schemas:
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        email:
          type: string
          minLength: 1
        phoneNumber:
          type: string
          maxLength: 15
          minLength: 0
        address:
          type: string
          maxLength: 255
          minLength: 0
        bio:
          type: string
        avatarUrl:
          type: string
        preferredLanguage:
          type: string
          enum:
          - ENGLISH
          - FRENCH
          - GERMAN
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
        contactOptions:
          type: array
          items:
            type: string
            enum:
            - EMAIL
            - PHONE
            - WHATSAPP
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        supabaseId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        contactOptions:
          type: array
          items:
            type: string
            enum:
            - EMAIL
            - PHONE
            - WHATSAPP
        phoneNumber:
          type: string
        address:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        preferredLanguage:
          type: string
          enum:
          - ENGLISH
          - FRENCH
          - GERMAN
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
        active:
          type: boolean
      required:
      - createdAt
      - email
      - id
      - name
      - supabaseId
    ItemRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 0
        description:
          type: string
          minLength: 1
        condition:
          type: string
          enum:
          - NEW
          - LIKE_NEW
          - USED
          - DEFECTIVE
        address:
          type: string
          minLength: 1
        isPickupPossible:
          type: boolean
        isShippingPossible:
          type: boolean
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: string
          format: uuid
      required:
      - categoryId
      - condition
      - imageUrls
      - isPickupPossible
      - isShippingPossible
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        parentId:
          type: string
          format: uuid
        parentTitle:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        active:
          type: boolean
      required:
      - active
      - createdAt
      - description
      - id
      - imageUrl
      - name
    ItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        condition:
          type: string
          enum:
          - NEW
          - LIKE_NEW
          - USED
          - DEFECTIVE
        address:
          type: string
        isPickupPossible:
          type: boolean
        isShippingPossible:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        imageUrls:
          type: array
          items:
            type: string
        itemType:
          type: string
          enum:
          - INTERNAL
          - EXTERNAL
        externalUrl:
          type: string
        status:
          type: string
          enum:
          - AVAILABLE
          - TAKEN
        active:
          type: boolean
        owner:
          $ref: "#/components/schemas/MinimalUserResponse"
        category:
          $ref: "#/components/schemas/CategoryResponse"
        sourceExternalPlatform:
          type: string
          enum:
          - ANIBIS
      required:
      - active
      - address
      - category
      - condition
      - createdAt
      - description
      - id
      - imageUrls
      - isPickupPossible
      - isShippingPossible
      - itemType
      - owner
      - status
      - title
    MinimalUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatarUrl:
          type: string
      required:
      - id
      - name
    CreateUserRequest:
      type: object
      properties:
        supabaseId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
          minLength: 0
        email:
          type: string
          minLength: 1
        authProvider:
          type: string
          enum:
          - SUPBASE
          - GOOGLE
          - FACEBOOK
        avatarUrl:
          type: string
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
      required:
      - authProvider
      - supabaseId
...
    PageItemResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ItemResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...