# src/lib/generated-api/models/MinimalUserResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MinimalUserResponse = {
    id: string;
    name: string;
    avatarUrl?: string;
};